code 09: Document everything related to the Hair Color catalog.
