+++ src/main/java/com/project/app_dog_back/domain/service/IColorPeloService.java
@@ -1,6 +1,6 @@
 package com.project.app_dog_back.domain.service;
 
-import com.project.app_dog_back.application.dto.ColorPeloDto;
+import com.project.app_dog_back.application.dto.catalog.ColorPeloDto;
 import com.project.app_dog_back.domain.model.catalog.ColorPelo;
 
 import java.util.List;
